StatusMessage: separate StatusIcon into its own component
